Remove shoot tactic. Move kick.py to tactic root directory
